Move `Channel::start_transfer` to `Transfer`
I'm going to add more methods that need to access channel registers
there anyway, and the start method seems to fit there better. Plus, all
the compiler fence stuff is now concentrated in `Channel`.
